dielectric, camera work

+++ src/gkk_material.cu
@@ -69,5 +69,5 @@ __device__ float schlick(float cosine, float ref_idx)
 {
   float r0 = (1.0f - ref_idx)/(1.0f + ref_idx);
   r0 = r0*r0;
-  return r0 + (1.0f - r0)*std::pow((1.0f - cosine), 5.0f);
+  return r0 + (1.0f - r0)*powf((1.0f - cosine), 5.0f);
 }
